requirements.txt
python-dotenv==1.0.1
discord.py==2.4.0
aiohttp~=3.9.5
pandas~=2.2.2
pandas_ta==0.3.14b0
numpy==1.26.4